feat: implementa middleware de autenticação e otimiza controllers

+++ src/controllers/tweetController.js
@@ -1,15 +1,9 @@
-import { tweetsCollection, usersCollection } from "../config/database.js";
+import { tweetsCollection } from "../config/database.js";
 import { ObjectId } from "mongodb";
 
 export async function createTweet(req, res) {
   try {
     const { username, tweet } = req.body;
-    
-    // Verifica se o usuário existe
-    const userExists = await usersCollection.findOne({ username });
-    if (!userExists) {
-      return res.status(401).send("Usuário não autorizado");
-    }
     
     await tweetsCollection.insertOne({
       username,
@@ -30,17 +24,32 @@ export async function getTweets(req, res) {
       .sort({ _id: -1 }) // Ordena do mais recente para o mais antigo
       .toArray();
     
-    // Adiciona avatar para cada tweet
-    const tweetsWithAvatar = [];
-    for (const tweet of tweets) {
-      const user = await usersCollection.findOne({ username: tweet.username });
-      tweetsWithAvatar.push({
-        ...tweet,
-        avatar: user.avatar
-      });
-    }
+    // Usar o mongoDB aggregation para juntar com a coleção de usuários e obter o avatar
+    const tweetsWithAvatars = await tweetsCollection.aggregate([
+      {
+        $lookup: {
+          from: "users",
+          localField: "username",
+          foreignField: "username",
+          as: "user"
+        }
+      },
+      {
+        $addFields: {
+          avatar: { $arrayElemAt: ["$user.avatar", 0] }
+        }
+      },
+      {
+        $project: {
+          user: 0
+        }
+      },
+      {
+        $sort: { _id: -1 }
+      }
+    ]).toArray();
     
-    return res.status(200).send(tweetsWithAvatar);
+    return res.status(200).send(tweetsWithAvatars);
   } catch (error) {
     console.error(error);
     return res.status(500).send("Erro interno do servidor");
@@ -50,7 +59,7 @@ export async function getTweets(req, res) {
 export async function updateTweet(req, res) {
   try {
     const { id } = req.params;
-    const { tweet } = req.body;
+    const { tweet, username } = req.body;
     
     // Verifica se o tweet existe
     const tweetExists = await tweetsCollection.findOne({ _id: new ObjectId(id) });
@@ -59,7 +68,7 @@ export async function updateTweet(req, res) {
     }
     
     // Verifica se o usuário é o dono do tweet
-    if (tweetExists.username !== req.body.username) {
+    if (tweetExists.username !== username) {
       return res.status(401).send("Você não tem permissão para editar este tweet");
     }
     
@@ -78,11 +87,17 @@ export async function updateTweet(req, res) {
 export async function deleteTweet(req, res) {
   try {
     const { id } = req.params;
+    const { username } = req.body;
     
     // Verifica se o tweet existe
     const tweetExists = await tweetsCollection.findOne({ _id: new ObjectId(id) });
     if (!tweetExists) {
       return res.status(404).send("Tweet não encontrado");
+    }
+    
+    // Verifica se o usuário é o dono do tweet
+    if (tweetExists.username !== username) {
+      return res.status(401).send("Você não tem permissão para deletar este tweet");
     }
     
     await tweetsCollection.deleteOne({ _id: new ObjectId(id) });
